feat: Parallel Compact

+++ file/file.go
@@ -3,7 +3,7 @@ package file
 import "io"
 
 type Options struct {
-	FID      uint32
+	FID      uint64
 	FileName string
 	Dir      string
 	Path     string
